feat: add joint limit calibration to controller
allows the arm to be constrained to a given location for safety.

The user will manually move the the arm through its range of motion and
the motors will not be commanded to go outside of it. I've been having
some issues with it spazzing out back and forth from overflow due to a
bad middle position set.